Move a big part of plugin.py to new pluginmanager.py
